move success message to parent component
